requirements_docker.txt
numpy==1.16.2
Pillow==5.4.1
pygame==2.0.1
pytest==6.2.3
PyYAML==5.4.1
scipy
filterpy
